[REM-103] refactor: clean up MasterApp and RunnerServices
- MasterApp: removed unused properties and setters for Celery, Quart, Server, ASGIApp, and Worker.
- RunnerServices: added properties and setters for Celery, Quart, Server, ASGIApp, and Worker; implemented start and stop methods for server management.
